SC-23 Add authorization check in deleteElevatorController

+++ src/controllers/elevator/delete/deleteElevatorController.ts
@@ -5,16 +5,23 @@ import { Result } from "../../../core/logic/Result";
 import IDeleteElevatorController from "../../IControllers/elevator/delete/IDeleteElevatorController";
 import IDeleteElevatorService from "../../../services/IServices/elevator/delete/IDeleteElevatorService";
 import IEditElevatorDTO from "../../../dto/elevator/IEditElevatorDTO";
+import { IAuthService } from "../../../services/IServices/auth/IAuthService";
 
 @Service()
 export default class DeleteElevatorController implements IDeleteElevatorController {
 
     constructor(
-        @Inject(config.services.deleteElevator.name) private service: IDeleteElevatorService
+        @Inject(config.services.deleteElevator.name) private service: IDeleteElevatorService,
+        @Inject(config.services.auth.name) private authService: IAuthService
     )
     {}
 
     public async deleteElevator(req: Request, res: Response, next: NextFunction) {
+        //@ts-ignore
+        let userRole = req.userRole;
+        if(!this.authService.validatePermission(userRole, ["CampusManager"])){
+            return res.status(401).send("Unauthorized");
+        }
         try {
             const elevatorOrError = await this.service.deleteElevator(req.body as IEditElevatorDTO) as Result<boolean>
